LayoutTests/canvas/philip/tests/2d.imageData.put.dirty.outside-expected.txt

< [index] >
2d.imageData.put.dirty.outside
putImageData() handles dirty rectangles outside the canvas correctly
References: 2d.imageData.put.normal
Actual output:
Expected output:

Passed
